% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colar.R
\name{colar}
\alias{colar}
\title{Cola x e y}
\usage{
colar(x, y)
}
\arguments{
\item{x}{primeiro parametro}

\item{y}{segundo parametro}
}
\value{
x e y colados
}
\description{
Recebe x e y e retorna eles colados
}
